update readme
Following the interview this morning, I decided to update this readme.

The candidate jumped straight on the readme section to get informations but the information is more a guide for the interviewers.

I decided to  create a readme to help the candidate.

The task 2 and 3 were removed from this readme to not scary candidates about the number of the task he has to complete in 40 minutes.

If the candidate finishes the task 1 the interviewer than can conduct the rest of the tests.